requirements.txt
aiogram==3.10.0
peewee==3.17.5
python-dotenv==1.0.1